requirements.txt
fastapi
uvicorn[standard]
transformers
sentence-transformers
torch
huggingface_hub
faiss-cpu
google-generativeai
numpy
python-dotenv
